refactor and implement react-hooks-form
