+++ client/src/features/users/Auth/SignUpNonProfit/Classifications.ts
@@ -1,0 +1,7 @@
+export const classifications: { value: string; text: string }[] = [
+  { value: 'charitable', text: 'Charitable Organization' },
+  { value: 'religious', text: 'Religious Organization' },
+  { value: 'private', text: 'Private Foundation' },
+  { value: 'political', text: 'Political Organizations' },
+  { value: 'other', text: 'Other' },
+];
