+++ client/src/features/users/Auth/SignUpNonProfit/FormData.ts
@@ -1,0 +1,18 @@
+export interface FormData {
+  name: string;
+  doing_business_as: string;
+  city: string;
+  state: string;
+  ein: string;
+  description: string;
+  website: string;
+  address: string;
+  phone: string;
+  nonprofit_classification: string;
+  firstName: string;
+  last_name: string;
+  email: string;
+  password: string;
+  confirmPassword: string;
+  accept_terms: boolean;
+}
