+++ client/src/features/users/Auth/SignUpNonProfit/styles.ts
@@ -1,0 +1,24 @@
+import { Theme } from '@mui/material';
+import makeStyles from '@mui/styles/makeStyles';
+import { placeholderImg } from '../../../../assets/temp';
+
+export const useStyles = makeStyles((theme: Theme) => ({
+  sideImg: {
+    backgroundImage: `url("${placeholderImg}")`,
+    backgroundSize: 'cover',
+    backgroundPosition: 'center center',
+    backgroundRepeat: 'no-repeat',
+  },
+  signUpContainer: {
+    margin: theme.spacing(5),
+  },
+  button: {
+    borderRadius: 0,
+    height: 62,
+    textTransform: 'none',
+    backgroundColor: '#C4C4C4',
+    color: 'white',
+  },
+  header: { fontWeight: 'bold' },
+  arrow: { cursor: 'pointer' },
+}));
